perf(web): convert App.svelte static imports to dynamic import() + refactor(web): split Settings.svelte into sub-components
